more consistent inline function definition for ArrowCharView()
